AG-17996 - port consent fields from ag-grid
Ports the AG-17996 consent field changes across from ag-grid
(ag-grid#14646, merged as a8e0dc175a9). The ag-website-shared files are
copied manually rather than via git subrepo, and are byte-identical to the
merged ag-grid versions.

Adds three web-to-lead consent checkboxes to the trial licence and contact
forms:

- Data Processing Consent - required to submit
- Email Marketing Consent - optional
- Email Tracking Consent - optional, revealed by a France/Italy
  self-declaration checkbox that stands in for IP geolocation

Also consolidates the trial licence form: the emailOnly/original variants
were never rendered because no call site passed a type to the dispatcher, so
the allFields implementation is promoted to TrialLicenceForm and the
dispatcher plus DEFAULT_TRIAL_LICENSE_FORM/TrialLicenceFormType are removed.
