+++ external/ag-website-shared/src/components/contact-form/ContactForm.module.scss
@@ -53,15 +53,10 @@
     font-family: inherit;
 }
 
-.privacyMessage {
-    font-size: var(--text-fs-xs);
-    line-height: var(--text-lh-xs);
-    color: color-mix(in srgb, var(--color-text-secondary), var(--color-bg-primary) 30%);
-    margin: 0;
-
-    a {
-        font-weight: var(--text-regular);
-    }
+.consents {
+    display: flex;
+    flex-direction: column;
+    gap: $spacing-size-2;
 }
 
 .tertiaryButton {
@@ -73,6 +68,8 @@
 }
 
 .errorContainer {
+    display: flex;
+    align-items: flex-start;
     margin-top: 0.25rem;
 
     &:empty {
