+++ packages/ag-charts-website/src/constants.ts
@@ -160,8 +160,6 @@ export const EXAMPLE_STYLE_FILE_NAME = 'ag-example-styles.css';
 export const DEBUG_SCRIPT_FILE_NAME = 'ag-charts-debug.js';
 
 export const ZI_FORM_ID = 'aad0527d-5af6-4263-8dcd-60f3ac998d5d';
-export type TrialLicenceFormType = 'emailOnly' | 'allFields' | 'original';
-export const DEFAULT_TRIAL_LICENSE_FORM: TrialLicenceFormType = 'allFields';
 
 // Google Tag Manager
 export const PUBLIC_GTM_ID = import.meta.env?.PUBLIC_GTM_ID;
